Added files via upload
This is the latest version I have as of 4/11
It works sort of, I'm not sure if the results are completely
correct.  But, at this point I think all the functions are working
properly.  I just probably need to clean up the code a bit.
I'm not sure how to test for whether the program is really
working properly or not.